[hotfix][table-api-java] Moved QueryOperation utilities to o.a.f.t.operations.utils
